Refactor extract components/models into separate files
